* New command 'S' to manage sections   - Change file.baddr depending on seek     > S                ; list sections\n");     > S*               ; list sections (in radare commands\n");     > S=               ; list sections (in visual)\n");     > S 4096 0x80000 rwx section_text  @ 0x8048000 ; adds new section\n");     > S 4096 0x80000   ; 4KB of section at current seek with base 0x.\n");     > S 10K @ 0x300    ; create 10K section at 0x300\n");     > S -0x300         ; remove this section definition\n");     > Sc rwx _text     ; add comment to the current section\n");     > Sb 0x100000      ; change base address\n");     > St 0x500         ; set end of section at this address\n");     > Sf 0x100         ; set from address of the current section\n"); * sS - seek to section index   - Fixes in the seek command for undo history and relative seeks   - Use undo push in '>' and '<' commands * Fix 64 bit bug in af (fixed function sizes)   - Added get0word() and set0word() in utils.c

+++ src/analyze.c
@@ -732,7 +732,7 @@ int analyze_function(int recursive, int report)
 	u64 end  = 0;
 	int i, inc = 0;
 	u64 to;
-	int len = 0;
+	u64 len = 0;
 	int ref;
 	int ncalls = 0;
 	int nrefs = 0;
